docs: add Discuss
Discuss is a forum like Reddit in dark theme. 

note: Hi I'm still pretty much a noob and still learning a lot of things. This is my first ever web app that I developed with Next.js + TS, and the first to be deployed on the cloud. I know it's still kinda scuffed and laggy and I'm still trying to optimize it, but it would be cool if this app could be featured here as this is my first attempt to contributing to open source. You guys made an awesome stack and I learned a lot about best practices especially watching Theo's videos and stream.