docs: Mount DATABASE_URL as secret instead of ARGS
Hi everyone.

Using DATABASE_URL as args and push the image in a public registry can lead to expose database sensitive information.

An attacker can easily inspect the docker image and see username and password if they are part of the DATABASE_URL.

So instead of doing that, I suggest mounting DATABASE_URL as a secret if we need to render pages on build time.
Using ENVS is great but in my case, I had to retrieve some db info on build time for SSG