print error.cause if not a ZodError
Refactor the errorFormatter function to handle error.cause more cleanly. 

If error.cause is undefined, it will not be included in the output. 
If it is an instance of ZodError, it will be included in the zodError property of the output, and if it is not, it will be included in the cause property of the output.