ci: :zap: Add /cli change validation to CI workflow
Added a new step in the CI workflow to check for changes in the /cli directory. If there are any changes, it will validate the changeset by running `pnpm changeset status --since origin/next`. This ensures that changes made to /cli are properly tracked and validated.

Should allow for changes to www and docs to be free from ci errors around changeset